refactor: rename and move line_sentence_handler
